[DDW-1147] Separate steps definitions regarding to features belongs to and tests improvements
